scripts/requirements.txt
bencode-python3
cbor
colorlog
mako
pip
plumbum >= 1.6.3
psutil
pygments
python-dateutil
typing;python_version<"3.5"
scan-build
pyyaml
toml
